@@ -1,6 +1,6 @@
 package world.best.musicplayer.activity;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.app.Activity;
 import android.database.Cursor;
 import android.os.Bundle;
