@@ -3,7 +3,8 @@ package world.best.musicplayer.factories;
 import android.content.Context;
 import android.net.Uri;
 import android.provider.MediaStore;
-import android.support.v4.content.CursorLoader;
+
+import androidx.loader.content.CursorLoader;
 
 /**
  * This is class is use for Create various loaders for loading data from Media Store.
