@@ -1,10 +1,11 @@
 package world.best.musicplayer.adapters;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
-import android.widget.TextView;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
+import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import world.best.musicplayer.R;
 
