@@ -1,8 +1,9 @@
 package world.best.musicplayer.adapters;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import world.best.musicplayer.R;
 
