@@ -2,13 +2,15 @@ package world.best.musicplayer.viewholder;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.view.View.OnClickListener;
+import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
+
 import world.best.musicplayer.R;
 import world.best.musicplayer.activity.TagActivity;
 import world.best.musicplayer.utils.Constants;
-import android.widget.TextView;
 
 public class AutoTagViewHolder extends RecyclerView.ViewHolder {
 
